+++ Sources/BaseUtility/SystemSettings/macOS/SettingsPane.swift
